separate python components
