+++ setup.py
@@ -2,7 +2,7 @@ from setuptools import setup, Extension
 
 bpipe = Extension(
     name="bpipe.core",  # <- matches your import path!
-    sources=["bpipe/core.c"],
+    sources=["bpipe/core.c", "bpipe/core_python.c"],
     extra_compile_args=["-g", "-O0"],
     extra_link_args=["-g"],
 )
